requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aioshutil==1.5
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
anthropic==0.57.1
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
blinker==1.9.0
boto3==1.39.4
botocore==1.39.4
cachetools==6.1.0
certifi==2025.7.9
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cryptography==45.0.5
dataclasses-json==0.6.7
distro==1.9.0
et-xmlfile==2.0.0
exceptiongroup==1.3.0
fastapi==0.116.0
filelock==3.18.0
frozenlist==1.7.0
fsspec==2025.5.1
gitdb==4.0.12
gitpython==3.1.44
greenlet==3.2.3
grpcio==1.73.1
h11==0.16.0
h2==4.2.0
hf-xet==1.1.5
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.33.2
hyperframe==6.1.0
idna==3.10
importlib-metadata==8.7.0
jieba==0.42.1
jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
langchain==0.3.26
langchain-community==0.3.27
langchain-core==0.3.68
langchain-openai==0.3.27
langchain-qdrant==0.2.0
langchain-text-splitters==0.3.8
langsmith==0.4.5
litellm==1.74.1
lxml==6.0.0
markupsafe==3.0.2
marshmallow==3.26.1
multidict==6.6.3
mypy-extensions==1.1.0
narwhals==1.46.0
numpy==2.3.1
openai==1.95.0
openpyxl==3.1.5
orjson==3.10.18
packaging==24.2
pandas==2.3.1
pdf2image==1.17.0
pdfminer-six==20250506
pdfplumber==0.11.7
pillow==11.3.0
plotly==6.2.0
portalocker==2.10.1
propcache==0.3.2
protobuf==6.31.1
py-zerox==0.0.7
pyarrow==20.0.0
pycparser==2.22
pydantic==2.11.7
pydantic-core==2.33.2
pydantic-settings==2.10.1
pydeck==0.9.1
pymupdf==1.26.3
pymupdf4llm==0.0.26
pypdf2==3.0.1
pypdfium2==4.30.1
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
python-multipart==0.0.20
python-pptx==1.0.2
pytz==2025.2
pyyaml==6.0.2
qdrant-client==1.14.3
referencing==0.36.2
regex==2024.11.6
reportlab==4.4.2
requests==2.32.4
requests-toolbelt==1.0.0
rpds-py==0.26.0
s3transfer==0.13.0
scipy==1.16.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sqlalchemy==2.0.41
starlette==0.46.2
streamlit==1.46.1
tenacity==9.1.2
tiktoken==0.9.0
tokenizers==0.21.2
toml==0.10.2
tornado==6.5.1
tqdm==4.67.1
typing-extensions==4.14.1
typing-inspect==0.9.0
typing-inspection==0.4.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
watchdog==6.0.0
xlrd==2.0.2
xlsxwriter==3.2.5
yarl==1.20.1
zipp==3.23.0
zstandard==0.23.0
